requirements.txt
blinker==1.4
certifi==2021.5.30
cffi==1.14.5
cryptography==3.4.7
fpdf==1.7.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.1
img2pdf==0.4.1
kaitaistruct==0.9
lxml==4.6.5
Pillow==9.0.1
pyasn1==0.4.8
pycparser==2.20
pydivert==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
selenium==3.141.0
selenium-wire==4.3.1
six==1.16.0
urllib3==1.26.5
wsproto==1.0.0
pikepdf==2.12.2
